Fiddling a bit, and I think we're getting really interesting results on the Netflix set now.
Made extract_key_users.pl allow the dropping of less-active users.  This made the clusters look better by a lot.

Gave score.cpp a parameter, popularity_reduction.  At 1, it scores a movie solely on how nicely-distributed-in-small-clusters its votes are; more votes are a liability if they don't affect the base score enough.  At 0, it has no effect.

Oddly, it seems to have strongly nonlinear behavior.  At popularity_reduction=0.68, only 2 of the top 16 movies have fewer than 50 votes, the rest have over 200.  At popularity_reduction=0.70, 15 of the top 16 have fewer than 50 votes, the remaining one is The Shawshank Redemption, with 384.

Looking at the low-vote movies that score well, it appears (subjectively) that they are obscure films that are critically acclaimed, as opposed to movies that were low-scoring because they were bad.  This is a really cool result -- our algorithm knows nothing but whether a user gave it 4+ stars; without knowing who watched the movie, but didn't like it, it was able to recover the difference between "bad" and "unpopular".