Okay, on these settings, the model's actual results are bonkers, but look at this: > rmsep(lm(ebert ~ mc, data=test)) [1] 0.6166564 > > rmsep(lm(ebert ~ nf, data=test)) [1] 0.8608464 > rmsep(lm(ebert ~ nf+yb, data=test)) [1] 0.7664548 > rmsep(lm(ebert ~ yb, data=test)) [1] 0.9180437
yb is driving the results down a significant portion of the way to
metacritic's predictive power.  This is interesting.  Again, if you
look at the actual results, you'll see that the best movie of all time
is by Prince ("Purple Rain"?), and the second-best film is a
"medicore" action film with a 9% rating at Rotten Tomatoes.  (Hm.  I
need to throw some trashy movies into the dataset.)  But it seems like
it's picking up something.  No idea what.